+++ src/web/views/component/social-links.blade.php
@@ -1,92 +1,56 @@
-<ul class="ui-list">
+<div class="hstack gap-1 fs-5">
     @if($public['social_url_facebook'])
-        <li class="ui-list-item">
-            <a
-                    href="{{$public['social_url_facebook']}}"
-                    target="_blank"
-                    class="ui-list-link"
-            >
-                <img
-                        src="/assets/social/facebook.svg"
-                        alt="Facebook"
-                        style="max-width: 28px;height: auto;"
-                />
-            </a>
-        </li>
+        <a
+                href="{{$public['social_url_facebook']}}"
+                target="_blank"
+                class="ui-list-link"
+        >
+            <i class="icon bi-facebook"></i>
+        </a>
     @endif
     @if($public['social_url_twitter'])
-        <li class="ui-list-item">
             <a
                     href="{{$public['social_url_twitter']}}"
                     target="_blank"
                     class="ui-list-link"
             >
-                <img
-                        src="/assets/social/twitter.svg"
-                        alt="Twitter"
-                        style="max-width: 28px;height: auto;"
-                />
+                <i class="icon bi-twitter"></i>
             </a>
-        </li>
     @endif
     @if($public['social_url_linkedin'])
-        <li class="ui-list-item">
             <a
                     href="{{$public['social_url_linkedin']}}"
                     target="_blank"
                     class="ui-list-link"
             >
-                <img
-                        src="/assets/social/linkedin.svg"
-                        alt="LinkedIn"
-                        style="max-width: 28px;height: auto;"
-                />
+                <i class="icon bi-linkedin"></i>
             </a>
-        </li>
     @endif
     @if($public['social_url_youtube'])
-        <li class="ui-list-item">
             <a
                     href="{{$public['social_url_youtube']}}"
                     target="_blank"
                     class="ui-list-link"
             >
-                <img
-                        src="/assets/social/youtube.svg"
-                        alt="YouTube"
-                        style="max-width: 28px;height: auto;"
-                />
+                <i class="icon bi-youtube"></i>
             </a>
-        </li>
     @endif
     @if($public['social_url_twitch'])
-        <li class="ui-list-item">
             <a
                     href="{{$public['social_url_twitch']}}"
                     target="_blank"
                     class="ui-list-link"
             >
-                <img
-                        src="/assets/social/twitch.svg"
-                        alt="Twitch"
-                        style="max-width: 28px;height: auto;"
-                />
+                <i class="icon bi-twitch"></i>
             </a>
-        </li>
     @endif
     @if($public['social_url_github'])
-        <li class="ui-list-item">
             <a
                     href="{{$public['social_url_github']}}"
                     target="_blank"
                     class="ui-list-link"
             >
-                <img
-                        src="/assets/social/github.svg"
-                        alt="GitHub"
-                        style="max-width: 28px;height: auto;"
-                />
+                <i class="icon bi-github"></i>
             </a>
-        </li>
     @endif
-</ul>
+</div>