+++ src/web/www/index.php
@@ -10,6 +10,7 @@ $view = new \core\module\web\View;
 
 $token = $as -> start_app_session();
 $meta = $view -> get_meta();
+$bs_token = 'sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p';
 ?>
 <!doctype html>
 <html lang="<?=($meta['lang']) ?>">
@@ -22,24 +23,16 @@ $meta = $view -> get_meta();
 	<meta name="robots" content="<?=($meta['robots']) ?>" />
 	<meta name="og:url" content="<?=($meta['url']) ?>" />
 
-	<link
-		href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-		rel="stylesheet"
-		integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-		crossorigin="anonymous"
-	/>
-	<link
-		rel="stylesheet"
-		href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css"
-	/>
+	<link rel="manifest" href="./manifest.json" />
+	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" />
 	<link href="<?= (WEB_DOCUMENT['styles']) ?>" rel="stylesheet" />
 
 	<script>
 		window.APP_ENV = window.APP_ENV || '<?=(ENV)?>';
 		window.APP_TIMESTAMP = window.APP_TIMESTAMP || '<?=(TIMESTAMP)?>';
 		window.APP_TOKEN = window.APP_TOKEN || '<?=($token)?>';
-		window.MEMBER_TOKEN = window.MEMBER_TOKEN || '<?=($as -> get_member_token())?>';
 		window.APP_LANG = window.APP_LANG || '<?=($meta['lang'])?>';
+		window.MEMBER_TOKEN = window.MEMBER_TOKEN || '<?=($as -> get_member_token())?>';
 	</script>
 </head>
 <body
@@ -52,11 +45,7 @@ $meta = $view -> get_meta();
 	>
 			<?php $view -> render() ?>
 	</div>
-	<script
-		src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-		integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-		crossorigin="anonymous"
-	></script>
+	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="<?=($bs_token)?>" crossorigin="anonymous"></script>
 	<script src="<?= (WEB_DOCUMENT['scripts']) ?>"></script>
 </body>
 </html>